Move logic to DynamicAgent
